Refactor filtering strategy functions and move to factory module
